+++ packages/sdk/src/models/base.ts
@@ -344,6 +344,7 @@ class Base extends AbstractModel<BaseData, WatchableBaseKey> {
      * @param name name for the table. must be case-insensitive unique
      * @param fields array of fields to create in the table
      *
+     * @example
      * ```js
      * const createTableCheckResult = base.unstable_checkPermissionsForCreateTable();
      *
@@ -390,6 +391,7 @@ class Base extends AbstractModel<BaseData, WatchableBaseKey> {
      * @param name name for the table. must be case-insensitive unique
      * @param fields array of fields to create in the table
      *
+     * @example
      * ```js
      * const canCreateTable = table.unstable_hasPermissionToCreateTable();
      *
@@ -421,7 +423,7 @@ class Base extends AbstractModel<BaseData, WatchableBaseKey> {
      * other specifics for certain field types.
      *
      * At least one field must be specified. The first field in the `fields` array will be used as
-     * the table's [primary field](https://support.airtable.com/hc/en-us/articles/202624179-The-Name-Field primary field)
+     * the table's [primary field](https://support.airtable.com/hc/en-us/articles/202624179-The-Name-Field)
      * and must be a supported primary field type. Fields must have case-insensitive unique names
      * within the table.
      *
@@ -432,8 +434,9 @@ class Base extends AbstractModel<BaseData, WatchableBaseKey> {
      * table in your block.
      *
      * @param name name for the table. must be case-insensitive unique
-     * @param fields array of fields to create in the table: see above for details.
+     * @param fields array of fields to create in the table: see below for details.
      *
+     * @example
      * ```js
      * async function createNewTable() {
      *     const name = 'My new table';