+++ packages/sdk/test/ui/choice_token.test.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import {mount} from 'enzyme';
+import {ChoiceToken} from '../../src/ui/ui';
+
+describe('ChoiceToken', () => {
+    it('renders outside of a blocks context', () => {
+        mount(<ChoiceToken choice={{id: 'test', name: 'Very important choice', color: 'blue'}} />);
+    });
+});